semly-backend/requirements.txt
asgiref==3.6.0
Django==3.2.11
django-cors-headers==3.13.0
djangorestframework==3.12.2
pytz==2022.7.1
sqlparse==0.4.3
